refactor: Restructure tests and update documentation for Interactions API
Test reorganization:
- Consolidate builder tests into builder_tests.rs (FunctionDeclaration + InteractionBuilder)
- Consolidate all API integration tests into interactions_api_tests.rs
- Remove redundant test files (expanded_e2e_tests.rs, function_calling_integration_tests.rs,
  function_declaration_builder_tests.rs, interaction_builder_edge_cases.rs, interactions_tests.rs)
- Keep macro_tests.rs as-is

Documentation updates:
- README.md: Complete rewrite for Interactions API (remove all GenerateContent references)
  - Updated examples for simple interactions, streaming, stateful conversations
  - Updated function calling examples with new FunctionResult pattern
  - Updated API reference for InteractionBuilder and InteractionResponse
- CLAUDE.md: Update test organization section to reflect new structure

Test coverage maintained at ~140 tests across:
- 46 unit tests in genai_client
- 28 unit tests in rust_genai
- 28 builder tests
- 25 integration tests (require API key)
- 9 macro tests
