+++ tests/builder_tests.rs
@@ -1,10 +1,256 @@
-// InteractionBuilder edge case tests
-//
-// These tests verify edge cases and error conditions for the InteractionBuilder.
-// Complementary to the unit tests in src/request_builder.rs
+//! Unit tests for builder patterns (no API key required)
+//!
+//! This file contains tests for:
+//! - FunctionDeclarationBuilder edge cases
+//! - InteractionBuilder edge cases and validation
 
-use genai_client::{InteractionContent, InteractionInput};
+use genai_client::{GenerationConfig, InteractionContent, InteractionInput};
 use rust_genai::{Client, FunctionDeclaration, WithFunctionCalling};
+use serde_json::json;
+
+// =============================================================================
+// FunctionDeclarationBuilder Tests
+// =============================================================================
+
+#[test]
+fn test_function_builder_with_empty_name() {
+    // Builder accepts empty names but logs a warning
+    let func = FunctionDeclaration::builder("")
+        .description("Test function")
+        .build();
+
+    assert_eq!(func.name(), "");
+    // Note: This will likely be rejected by the API, but builder allows it
+}
+
+#[test]
+fn test_function_builder_with_whitespace_only_name() {
+    let func = FunctionDeclaration::builder("   ")
+        .description("Test function")
+        .build();
+
+    assert_eq!(func.name(), "   ");
+}
+
+#[test]
+fn test_function_builder_parameter_overwrites_on_duplicate() {
+    // When the same parameter name is added twice, the second should overwrite
+    let func = FunctionDeclaration::builder("test_func")
+        .parameter(
+            "location",
+            json!({"type": "string", "description": "First"}),
+        )
+        .parameter(
+            "location",
+            json!({"type": "string", "description": "Second"}),
+        )
+        .build();
+
+    // Verify the second parameter overwrote the first
+    let location_desc = func
+        .parameters()
+        .properties()
+        .get("location")
+        .and_then(|l| l.get("description"))
+        .and_then(|d| d.as_str());
+
+    assert_eq!(location_desc, Some("Second"));
+}
+
+#[test]
+fn test_function_builder_required_non_existent_parameter() {
+    // Builder allows requiring parameters that don't exist but logs a warning
+    let func = FunctionDeclaration::builder("test_func")
+        .parameter("existing_param", json!({"type": "string"}))
+        .required(vec!["nonexistent_param".to_string()])
+        .build();
+
+    assert_eq!(func.parameters().required(), vec!["nonexistent_param"]);
+    // Note: This will likely cause API errors, but builder allows it
+}
+
+#[test]
+fn test_function_builder_method_order_independence() {
+    // Verify that calling methods in different orders produces identical results
+    let func1 = FunctionDeclaration::builder("test")
+        .description("A test function")
+        .parameter("param1", json!({"type": "string"}))
+        .required(vec!["param1".to_string()])
+        .build();
+
+    let func2 = FunctionDeclaration::builder("test")
+        .required(vec!["param1".to_string()])
+        .parameter("param1", json!({"type": "string"}))
+        .description("A test function")
+        .build();
+
+    // Compare serialized forms since FunctionDeclaration doesn't implement PartialEq
+    let json1 = serde_json::to_value(&func1).unwrap();
+    let json2 = serde_json::to_value(&func2).unwrap();
+
+    assert_eq!(json1, json2);
+}
+
+#[test]
+fn test_function_builder_with_no_parameters() {
+    let func = FunctionDeclaration::builder("no_params")
+        .description("Function with no parameters")
+        .build();
+
+    assert_eq!(func.parameters().type_(), "object");
+    assert!(func.parameters().properties().is_object());
+    assert!(func.parameters().required().is_empty());
+}
+
+#[test]
+fn test_function_builder_with_many_parameters() {
+    let mut builder = FunctionDeclaration::builder("many_params");
+
+    // Add 20 parameters
+    for i in 0..20 {
+        builder = builder.parameter(
+            &format!("param_{}", i),
+            json!({"type": "string", "description": format!("Parameter {}", i)}),
+        );
+    }
+
+    let func = builder.build();
+
+    // Verify all parameters were added
+    let properties = func.parameters().properties();
+
+    for i in 0..20 {
+        assert!(properties.get(format!("param_{}", i)).is_some());
+    }
+}
+
+#[test]
+fn test_function_builder_required_all_parameters() {
+    let func = FunctionDeclaration::builder("all_required")
+        .parameter("param1", json!({"type": "string"}))
+        .parameter("param2", json!({"type": "number"}))
+        .parameter("param3", json!({"type": "boolean"}))
+        .required(vec![
+            "param1".to_string(),
+            "param2".to_string(),
+            "param3".to_string(),
+        ])
+        .build();
+
+    assert_eq!(func.parameters().required().len(), 3);
+}
+
+#[test]
+fn test_function_builder_required_subset_of_parameters() {
+    let func = FunctionDeclaration::builder("partial_required")
+        .parameter("required_param", json!({"type": "string"}))
+        .parameter("optional_param", json!({"type": "string"}))
+        .required(vec!["required_param".to_string()])
+        .build();
+
+    assert_eq!(func.parameters().required().len(), 1);
+    assert_eq!(func.parameters().required()[0], "required_param");
+}
+
+#[test]
+fn test_function_builder_complex_nested_schema() {
+    let func = FunctionDeclaration::builder("nested_schema")
+        .parameter(
+            "complex_param",
+            json!({
+                "type": "object",
+                "properties": {
+                    "nested": {
+                        "type": "object",
+                        "properties": {
+                            "deep": {"type": "string"}
+                        }
+                    }
+                }
+            }),
+        )
+        .build();
+
+    // Verify nested structure is preserved
+    let param = func.parameters().properties().get("complex_param");
+
+    assert!(param.is_some());
+    assert!(param.unwrap().get("properties").is_some());
+}
+
+#[test]
+fn test_function_builder_with_array_parameter() {
+    let func = FunctionDeclaration::builder("array_param")
+        .parameter(
+            "items",
+            json!({
+                "type": "array",
+                "items": {"type": "string"}
+            }),
+        )
+        .build();
+
+    let param = func
+        .parameters()
+        .properties()
+        .get("items")
+        .and_then(|i| i.get("type"))
+        .and_then(|t| t.as_str());
+
+    assert_eq!(param, Some("array"));
+}
+
+#[test]
+fn test_function_builder_with_enum_values() {
+    let func = FunctionDeclaration::builder("enum_param")
+        .parameter(
+            "unit",
+            json!({
+                "type": "string",
+                "enum": ["celsius", "fahrenheit", "kelvin"]
+            }),
+        )
+        .build();
+
+    let enum_values = func
+        .parameters()
+        .properties()
+        .get("unit")
+        .and_then(|u| u.get("enum"))
+        .and_then(|e| e.as_array());
+
+    assert_eq!(enum_values.unwrap().len(), 3);
+}
+
+#[test]
+fn test_function_builder_description_can_be_empty() {
+    let func = FunctionDeclaration::builder("test").description("").build();
+
+    assert_eq!(func.description(), "");
+}
+
+#[test]
+fn test_function_builder_description_with_unicode() {
+    let func = FunctionDeclaration::builder("test")
+        .description("测试函数 with émojis 🎉")
+        .build();
+
+    assert!(func.description().contains("🎉"));
+}
+
+#[test]
+fn test_function_builder_very_long_description() {
+    let long_desc = "x".repeat(10000);
+    let func = FunctionDeclaration::builder("test")
+        .description(&long_desc)
+        .build();
+
+    assert_eq!(func.description().len(), 10000);
+}
+
+// =============================================================================
+// InteractionBuilder Tests
+// =============================================================================
 
 #[test]
 fn test_interaction_builder_with_complex_content_input() {
@@ -40,16 +286,11 @@ fn test_interaction_builder_with_complex_content_input() {
         }
         _ => panic!("Expected InteractionInput::Content variant"),
     }
-
-    // Builder accepts complex multi-part content without validation
-    // Actual structure validation happens during API request creation
-    // This test ensures the builder API supports heterogeneous content arrays
 }
 
 #[test]
 fn test_interaction_builder_with_both_model_and_agent_set() {
     // Test that setting both model AND agent is allowed (builder doesn't validate this)
-    // The validation happens at request creation time
     let client = Client::new("test-api-key".to_string(), None);
 
     let _builder = client
@@ -60,7 +301,6 @@ fn test_interaction_builder_with_both_model_and_agent_set() {
 
     // Builder allows setting both model and agent without validation
     // The API will reject this at request time, not during builder construction
-    // This verifies the builder follows a "fail late" validation strategy
 }
 
 #[test]
@@ -75,8 +315,6 @@ fn test_interaction_builder_with_very_long_text() {
         .with_text(&long_text);
 
     // Builder accepts large text inputs without size validation
-    // Size limits are enforced by the API, not the builder
-    // This test ensures no artificial limits are imposed client-side
 }
 
 #[test]
@@ -92,7 +330,6 @@ fn test_interaction_builder_with_unicode_and_emojis() {
         .with_text(unicode_text);
 
     // Builder handles Unicode and multi-byte characters correctly
-    // This verifies proper UTF-8 string handling without mangling
 }
 
 #[test]
@@ -106,15 +343,11 @@ fn test_interaction_builder_with_empty_text() {
         .with_text("");
 
     // Builder allows empty string inputs without validation
-    // The API determines whether empty text is acceptable for a given request
 }
 
 #[test]
 fn test_interaction_builder_with_multiple_functions() {
     // Test adding many functions
-    use rust_genai::FunctionDeclaration;
-    use serde_json::json;
-
     let client = Client::new("test-api-key".to_string(), None);
 
     let mut builder = client
@@ -133,7 +366,6 @@ fn test_interaction_builder_with_multiple_functions() {
     }
 
     // Builder accepts many function declarations without validation
-    // This verifies the builder can handle complex tool configurations
 }
 
 #[test]
@@ -141,7 +373,7 @@ fn test_interaction_builder_with_complex_generation_config() {
     // Test with generation config at boundary values
     let client = Client::new("test-api-key".to_string(), None);
 
-    let config = genai_client::GenerationConfig {
+    let config = GenerationConfig {
         temperature: Some(2.0),        // Max value
         max_output_tokens: Some(8192), // High value
         top_p: Some(1.0),              // Max value
@@ -156,14 +388,11 @@ fn test_interaction_builder_with_complex_generation_config() {
         .with_generation_config(config);
 
     // Builder accepts generation config with boundary values
-    // Parameter range validation is performed by the API, not the builder
 }
 
 #[test]
 fn test_interaction_builder_with_response_format_json_schema() {
     // Test with complex JSON schema for structured output
-    use serde_json::json;
-
     let client = Client::new("test-api-key".to_string(), None);
 
     let complex_schema = json!({
@@ -194,39 +423,6 @@ fn test_interaction_builder_with_response_format_json_schema() {
         .with_response_format(complex_schema);
 
     // Builder accepts complex nested JSON schemas without validation
-    // Schema correctness is validated by the API when making the request
-}
-
-#[tokio::test]
-#[ignore = "Requires API key and makes real HTTP request"]
-async fn test_interaction_builder_with_longer_prompt() {
-    // Test the builder with a longer prompt that should work
-    let Ok(api_key) = std::env::var("GEMINI_API_KEY") else {
-        println!("Skipping test_interaction_builder_with_longer_prompt: GEMINI_API_KEY not set.");
-        return;
-    };
-
-    let client = Client::new(api_key, None);
-
-    // Request a simple response using the builder
-    let response = client
-        .interaction()
-        .with_model("gemini-3-flash-preview")
-        .with_text("Count from 1 to 5")
-        .create()
-        .await;
-
-    assert!(
-        response.is_ok(),
-        "Interaction builder should successfully create interaction: {:?}",
-        response.err()
-    );
-
-    let interaction = response.unwrap();
-    assert!(!interaction.id.is_empty(), "Should have interaction ID");
-    assert!(!interaction.outputs.is_empty(), "Should have outputs");
-
-    println!("Interaction completed with ID: {}", interaction.id);
 }
 
 #[test]
@@ -238,7 +434,7 @@ fn test_interaction_builder_with_all_features_combined() {
         .description("Get weather")
         .build();
 
-    let config = genai_client::GenerationConfig {
+    let config = GenerationConfig {
         temperature: Some(0.7),
         max_output_tokens: Some(1024),
         top_p: Some(0.95),
@@ -258,7 +454,6 @@ fn test_interaction_builder_with_all_features_combined() {
         .with_store(false);
 
     // Builder supports combining all features without conflicts
-    // This tests that the builder API is fully composable
 }
 
 #[test]
@@ -279,7 +474,9 @@ fn test_interaction_builder_method_chaining() {
     // All methods should be chainable and later calls overwrite earlier values
 }
 
-// Validation error tests
+// =============================================================================
+// InteractionBuilder Validation Tests
+// =============================================================================
 
 #[test]
 fn test_interaction_builder_validation_missing_input() {
